Refrain from pop-up notifications on parsing errors when editor is focused on .git/machete
